Multicommodity simulator (#94)
Start of the multicommodity simulator workflow
The merrit order is applied using the ESDL marginal costs attribute. 
Deman profiles are always matched first before the maximisation of demand and minimisation of production goals are applied. It already includes the prioritisation of the electrolyzer.
Omitting the use of the TechnoEconomicMixin, to prevent the creation of unnecessary variables.