Multicommodity simulator storage (#119)
This merge commit includes the storages in the multicommodity simulator workflow. The goals are on the charging and discharging variables, with the discharging to be minimised and the charging to be maximised.
Test case on the storage are also implented.

Furthermore this commit also includes a testcase for the simulator where the headloss is included.