Sequential mc simulator (#125)
Implementation of sequential staged class and run function for the multicommodity simulator.  A defined windowsize is used to split up the problem in multiple staged optimizations. The sequential staged approach passes on the last timestep of the stage to the next stage as the initial condition, to ensure continuity of the timehorizon of the assets. 
The test checks also the condition in which the result between the staged and unstaged are the same and also the condition when they might be different.

---------

Co-authored-by: Rojer <jim.rojer@tno.nl>