166 adapt profiles for heating and cooling peak day seperately (#181)
The goal of the PR was to allow mesido (in the presence of cold demands) to discretize hourly the peak days for both maximum heat and cold demands. This was done mainly by developing the adapt_profiles.py script, where the time series are modified. A few tests were added to account for this new functionality. These were placed in the test_cold_demand.py test file.

---------

Co-authored-by: kobus.vanrooyen@tno.nl <kobus.vanrooyen@tno.nl>