Heat exchanger bypass (#218)
Heat exchanger bypass is added as an option, which allows for timesteps with a flowrate passing the heat exchanger on the primary or secondary side, but without any heat heat transfer, in case the temperature on the return and supply of one side are the same. The temperature on the primary side should then also be equal or smaller than the temperature on the secondary side.