Heat exchanger bypass allowed for variable temperatures (#239)
Updated the heatexchanger heat_to_discharge constraints such that is also working for the bypass when variable temperature are picked on either side of the heat exchanger. This includes a temporary fix for a big_m which is set to infinite when temperature options are available and source is connected with logical link.